Update r2pro network
I've been testing out your builds on my r2pro. I really appreciate the effort you've put in on it.

I'm not sure where the change occurred, but the network interface mapping changed. I've tested this edit and it works as expected.

In the latest builds there is also an "erspan0" adapter and eth1 shows up as a network switch. Not sure if those should be included in ucidef_set_interfaces_lan_wan, but thought it's worth mentioning.